Organize components
Spreads components into molecules, atoms, organism. This is just a first approach, which can be fine-tuned iteratively.
